5379: create mapper file (cleanup)
